+++ ios/Classes/PluginScriptsJS/ConsoleLogJS.swift
@@ -32,19 +32,25 @@ let CONSOLE_LOG_JS_SOURCE = """
     for (var k in oldLogs) {
         (function(oldLog) {
             console[oldLog.replace('console', '').toLowerCase()] = function() {
-                var message = '';
-                for (var i in arguments) {
-                    if (message == '') {
-                        message += arguments[i];
-                    }
-                    else {
-                        message += ' ' + arguments[i];
-                    }
-                }
-
-                var _windowId = \(WINDOW_ID_VARIABLE_JS_SOURCE);
-                window.webkit.messageHandlers[oldLog].postMessage({'message': message, '_windowId': _windowId});
                 oldLogs[oldLog].apply(null, arguments);
+                var args = arguments;
+
+                // on iOS, for some reason, accessing the arguments object synchronously can throw some errors, such as "TypeError"
+                // see https://github.com/pichillilorenzo/flutter_inappwebview/issues/776
+                setTimeout(function() {
+                    var message = '';
+                    for (var i in args) {
+                        if (message == '') {
+                            message += args[i];
+                        }
+                        else {
+                            message += ' ' + args[i];
+                        }
+                    }
+
+                    var _windowId = \(WINDOW_ID_VARIABLE_JS_SOURCE);
+                    window.webkit.messageHandlers[oldLog].postMessage({'message': message, '_windowId': _windowId});
+                });
             }
         })(k);
     }