+++ lib/src/in_app_webview/ios/in_app_webview_options.dart
@@ -211,9 +211,10 @@ class IOSInAppWebViewOptions
   ///**NOTE**: available on iOS 13.0+.
   bool applePayAPIEnabled;
 
-  ///Used in combination with [WebView.initialUrlRequest] (using the `file://` scheme), it represents the URL from which to read the web content.
+  ///Used in combination with [WebView.initialUrlRequest] or [WebView.initialData] (using the `file://` scheme), it represents the URL from which to read the web content.
   ///This URL must be a file-based URL (using the `file://` scheme).
-  ///Specify the same value as the [WebView.initialUrlRequest] parameter to prevent WebView from reading any other content.
+  ///Specify the same value as the [URLRequest.url] if you are using it with the [WebView.initialUrlRequest] parameter or
+  ///the [InAppWebViewInitialData.baseUrl] if you are using it with the [WebView.initialData] parameter to prevent WebView from reading any other content.
   ///Specify a directory to give WebView permission to read additional files in the specified directory.
   Uri? allowingReadAccessTo;