+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 5.3.1
+
+- Added support of `allowingReadAccessTo` iOS-specific WebView option for the WebView `initialData` parameter
+- Added `iosAllowingReadAccessTo` iOS-specific parameter to the `loadData` WebView method
+- Fixed "iOS webview showing blank page in specific URL" [#776](https://github.com/pichillilorenzo/flutter_inappwebview/issues/776)
+- Fixed "unable to access ApplicationDocumentsDirectory in real Ios devices" [#748](https://github.com/pichillilorenzo/flutter_inappwebview/issues/748)
+
 ## 5.3.0+1
 
 - Fixed "Android - Pull to refresh triggered when scrolling container inside a website" [#765](https://github.com/pichillilorenzo/flutter_inappwebview/issues/765)